+++ src/Resources/contao/languages/de/modules.php
@@ -14,8 +14,8 @@ $GLOBALS['TL_LANG']['MOD']['fernschach'] = array('Fernschach-Verwaltung', 'Ferns
 $GLOBALS['TL_LANG']['MOD']['fernschach-spieler'] = array('Spieler', 'Spieler verwalten');
 $GLOBALS['TL_LANG']['MOD']['fernschach-meldungen'] = array('Meldungen', 'Turnier-Anmeldungen verwalten');
 $GLOBALS['TL_LANG']['MOD']['fernschach-turnierkategorien'] = array('Turnierkategorien', 'Turnierkategorien und Turnierhauptklassen verwalten');
-$GLOBALS['TL_LANG']['MOD']['fernschach-turniere'] = array('Turniere', 'Turniere verwalten');
 $GLOBALS['TL_LANG']['MOD']['fernschach-mitgliederstatistik'] = array('Mitgliederstatistik', 'Mitgliederstatistiken verwalten und ausgeben');
+$GLOBALS['TL_LANG']['MOD']['fernschach-turniere'] = array('Turniere (Alpha-Version)', 'Turniere verwalten');
 
 /**
  * Frontend-Module