+++ src/Resources/contao/languages/de/tl_user_group.php
@@ -1,0 +1,48 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2005-2023 Leo Feyer
+ *
+ * @package   fernschach
+ * @author    Frank Hoppe
+ * @license   GNU/LPGL
+ * @copyright Frank Hoppe 2023
+ */
+
+/**
+ * Backend-Modul Übersetzungen
+ */
+
+$GLOBALS['TL_LANG']['tl_user_group']['fernschach_legend'] = 'Fernschach-Verwaltung';
+$GLOBALS['TL_LANG']['tl_user_group']['fernschach_spieler'] = array('Spielerrechte', 'Hier können Sie den Zugriff auf die Spieler festlegen.');
+$GLOBALS['TL_LANG']['tl_user_group']['fernschach_konto'] = array('Kontorechte', 'Hier können Sie den Zugriff auf das Konto der Spieler festlegen.');
+
+$GLOBALS['TL_LANG']['tl_user_group']['fernschach_spieler_optionen'] = array
+(
+	'import'     => 'Spieler importieren',
+	'export'     => 'Spieler exportieren',
+	'create'     => 'Spieler anlegen',
+	'edit'       => 'Spieler bearbeiten',
+	'delete'     => 'Spieler löschen',
+	'copy'       => 'Spieler kopieren',
+	'toggle'     => 'Spieler Veröffentlichen-Status setzen',
+	//'show'       => 'Spieler Infobox zeigen',
+	'fertig'     => 'Spieler Fertig-Status setzen',
+	'all'        => 'Mehrere Spieler bearbeiten',
+	'konto'      => 'Buchungen anzeigen',
+);
+
+$GLOBALS['TL_LANG']['tl_user_group']['fernschach_konto_optionen'] = array
+(
+	'import'     => 'Buchungen importieren',
+	'create'     => 'Buchung anlegen',
+	'edit'       => 'Buchung bearbeiten',
+	'delete'     => 'Buchung löschen',
+	'copy'       => 'Buchung kopieren',
+	'toggle'     => 'Buchung Veröffentlichen-Status setzen',
+	//'show'       => 'Buchung Infobox zeigen',
+	'fertig'     => 'Buchung Fertig-Status setzen',
+	'all'        => 'Mehrere Buchungen bearbeiten',
+);